Made return types compatible with mainstream

+++ msprintf.c
@@ -41,7 +41,7 @@
 #include "msprintf.h"
 
 FUN_FORMAT(printf, 3, 0)
-size_t vasnprintf(char **out, size_t lim, const char *fmt, va_list a) {
+int vasnprintf(char **out, size_t lim, const char *fmt, va_list a) {
   int len, chk;
   char *str;
   va_list x;
@@ -72,7 +72,7 @@ size_t vasnprintf(char **out, size_t lim, const char *fmt, va_list a) {
 }
 
 FUN_FORMAT(printf, 3, 4)
-size_t asnprintf(char **out, size_t limit, const char *fmt, ...) {
+int asnprintf(char **out, size_t limit, const char *fmt, ...) {
   size_t res;
   va_list a;
   va_start(a, fmt);
@@ -84,12 +84,12 @@ size_t asnprintf(char **out, size_t limit, const char *fmt, ...) {
 #ifndef MISSING_ASPRINTF
 
 FUN_FORMAT(printf, 2, 0)
-size_t vasprintf(char **out, const char *fmt, va_list a) {
+int vasprintf(char **out, const char *fmt, va_list a) {
   return vasnprintf(out, INT_MAX, fmt, a);
 }
 
 FUN_FORMAT(printf, 2, 3)
-size_t asprintf(char **out, const char *fmt, ...) {
+int asprintf(char **out, const char *fmt, ...) {
   size_t res;
   va_list a;
   va_start(a, fmt);
